Move ApproxEqualityConstraint into Constraints namespace
